refactor(components): [backtop] (#10151)
* refactor(components): [backtop]

* Extract logic code from component itself.
* Reorganize code for better readability.

* chore: remove handleScrollThrottled from being returned
