+++ packages/components/backtop/index.ts
@@ -6,3 +6,4 @@ export const ElBacktop = withInstall(Backtop)
 export default ElBacktop
 
 export * from './src/backtop'
+export type { BacktopInstance } from './src/instance'
