+++ packages/components/backtop/src/backtop.ts
@@ -1,5 +1,4 @@
 import type { ExtractPropTypes } from 'vue'
-import type Backtop from './backtop.vue'
 
 export const backtopProps = {
   visibilityHeight: {
@@ -25,5 +24,3 @@ export const backtopEmits = {
   click: (evt: MouseEvent) => evt instanceof MouseEvent,
 }
 export type BacktopEmits = typeof backtopEmits
-
-export type BacktopInstance = InstanceType<typeof Backtop>
