+++ packages/components/backtop/src/instance.ts
@@ -1,0 +1,3 @@
+import type Backtop from './backtop.vue'
+
+export type BacktopInstance = InstanceType<typeof Backtop>
